requirements.txt
torch==2.8.0
transformers==4.57.3
accelerate==1.10.1
tqdm==4.67.1
sentence-transformers==5.1.2
wandb==0.23.1
numpy==2.0.2
datasets==2.21.0
modelscope==1.33.0
addict==2.4.0
python-dotenv==1.2.1